Rename i16n -> i18n
This should be correct ;)
